+++ bets/views.py
@@ -150,7 +150,7 @@ class BetRanking(View):
       # TODO filter parmi les tournois
       match_bets        = MatchBet.objects.filter(player = user)
       tournament_bets   = TournamentBet.objects.filter(player = user).last()
-      score     = sum(int(b['points_won'] if b['points_won'] is not None else 0 ) for b in match_bets.values())
+      score     = sum(float(b['points_won'] if b['points_won'] is not None else 0 ) for b in match_bets.values())
       try:
         # if tournament_bets != Null
         scoref    = score + tournament_bets.points_won
@@ -160,8 +160,8 @@ class BetRanking(View):
       # find best score per prognosis
       best_score = 0
       for b in match_bets.values():
-        if best_score <= int(b['points_won'] if b['points_won'] is not None else 0):
-          best_score = int(b['points_won'] if b['points_won'] is not None else 0)
+        if best_score <= float(b['points_won'] if b['points_won'] is not None else 0):
+          best_score = float(b['points_won'] if b['points_won'] is not None else 0)
       rank.append( (user.username, score, best_score) )
     #rank.sort(key=lambda r: r[1], reverse=True)
     # https://wiki.python.org/moin/HowTo/Sorting
